Working version
Working version for the small test put (to test on bigger instance). Some bug fix such as a bad relocation and bad insertion due to false demand supplied.
About the python: Added method to provide the list of customers require by sarl program and the depot. Can also interpret and build solution based on the response of sarl.